refactor(LottoController.java) : 숫자 유틸 convert 메서드 컨트롤러 이동
