+++ docs/README.md
@@ -95,7 +95,7 @@
     - [x] 구매 횟수 만큼의 로또 번호 리스트를 생성한다.
 
 ### 로또 집합(Lottos)
-- 
+
 - [x] 일치 횟수를 만족하는 로또의 갯수를 구한다.
 
 ### 로또 컨트롤러
@@ -128,8 +128,3 @@
 - [x] 수익율을 구한다.
   - [x] 구매 구입 금액에서 당첨 금액을 차감한 후 구매 금액으로 나눈다.
   - [x] 소수점 두자리
-
-
-### 숫자 유틸
-
-- [x] 문자 배열을 숫자 ArrayList로 변환한다.
