+++ src/test/java/lotto/controller/LottoControllerTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -10,14 +10,14 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DisplayName("숫자유틸")
-class NumberUtilTest {
+@DisplayName("로또 컨트롤러")
+class LottoControllerTest {
 
     @DisplayName("문자 숫자 리스트 변환")
     @ParameterizedTest
     @MethodSource("convert")
     void convert(String[] text) {
-        assertThat(NumberUtil.convert(text)).containsExactly(1, 2, 3, 4, 5, 6);
+        assertThat(LottoController.convert(text)).containsExactly(1, 2, 3, 4, 5, 6);
     }
 
     static Stream<Arguments> convert() {
